fe: Add getResources and getTags functions which encapsulate the resource fetching code. Refactor ResourceList.jsx to get its resource data via these new functions.
Explanation:
The idea here behind these new functions (getResources and getTags) is to encapsulate
the resource fetching code.

Any code that needs resource data should call these functions instead of calling fetch().
That way, the code doesn't have to care where the resource data is coming from, just that it gets the data.

This is useful for because we can easily change how and where we're getting the resource data without having to reprogram other parts of the app. To put it specifically, even though the proxy server is not finished and usable yet, we can fake/mock its functionality, and work on other features as if it is finished.

Simply put:
Since (as of this commit) the proxy server is not finished, these functions return data stored in local files.
After the proxy server is finished, we can update these functions to fetch data from the proxy server instead.